[infra] Replace Debian10 GCE tasks with Ubuntu 24.04
Bug: b/380037738
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/927102
Commit-Queue: Eric Boren <borenet@google.com>

+++ tests/CoreBlittersTest.cpp
@@ -87,8 +87,8 @@ static void compare_mask_and_antiH(skiatest::Reporter* reporter,
             REPORT_FAILURE(reporter,
                            "blitAntiH was not the same for all pixels",
                            SkStringPrintf("background=%08x, paint=%08x, alpha=%d",
-                                          backgroundColor,
-                                          paintColor,
+                                          unsigned(backgroundColor),
+                                          unsigned(paintColor),
                                           alpha));
             return;
         }
@@ -96,8 +96,8 @@ static void compare_mask_and_antiH(skiatest::Reporter* reporter,
             REPORT_FAILURE(reporter,
                            "blitMask was not the same for all pixels",
                            SkStringPrintf("background=%08x, paint=%08x, alpha=%d",
-                                          backgroundColor,
-                                          paintColor,
+                                          unsigned(backgroundColor),
+                                          unsigned(paintColor),
                                           alpha));
             return;
         }
@@ -110,12 +110,12 @@ static void compare_mask_and_antiH(skiatest::Reporter* reporter,
                         "background=%08x, paint=%08x, alpha=%d, "
                         "blitAntiH=%08x, blitMask=%08x, "
                         "diff=%02x %02x %02x %02x",
-                        backgroundColor, paintColor, alpha,
-                        antiHColor, maskColor,
-                        abs((int)(SkColorGetA(antiHColor) - SkColorGetA(maskColor))),
-                        abs((int)(SkColorGetR(antiHColor) - SkColorGetR(maskColor))),
-                        abs((int)(SkColorGetG(antiHColor) - SkColorGetG(maskColor))),
-                        abs((int)(SkColorGetB(antiHColor) - SkColorGetB(maskColor))));
+                        unsigned(backgroundColor), unsigned(paintColor), alpha,
+                        unsigned(antiHColor), unsigned(maskColor),
+                        unsigned(abs((int)(SkColorGetA(antiHColor) - SkColorGetA(maskColor)))),
+                        unsigned(abs((int)(SkColorGetR(antiHColor) - SkColorGetR(maskColor)))),
+                        unsigned(abs((int)(SkColorGetG(antiHColor) - SkColorGetG(maskColor)))),
+                        unsigned(abs((int)(SkColorGetB(antiHColor) - SkColorGetB(maskColor)))));
     }
 }
